pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-lokalise"
authors = [
    {name = "Jonathan Perron", email = "contact@jonathanperron.fr"},
]
license = {text = "BSD 3"}
description = "Lokalise hook and operator for Airflow based on the Lokalise Python SDK"
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.7",
    "apache-airflow-providers-postgres>=5.8.0, <6",
    "python-lokalise-api>=2.1.0, <3",
]

[project.urls]
Source = "https://github.com/jonperron/airflow-provider-lokalise"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "lokalise_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "lokalise_provider.__version__"}

[tool.setuptools.packages.find]
exclude = ["*tests.*", "*tests"]
